Move all code from generate-node.js to generate.js, create package.json for script, update README.md
