+++ .github/workflows/build-grammars.yml
@@ -10,7 +10,7 @@ jobs:
         node-version: [18.x]
     steps:
     - uses: actions/checkout@v3
-    - run: node generate-node.js
+    - run: node generate.js
     - name: Upload a Build Artifact
       uses: actions/upload-artifact@v3.1.3
       with:
