+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "libprisma",
+  "version": "1.0.0",
+  "description": "This is a C++ porting of [prism.js](https://github.com/PrismJS/prism) library.\r The code depends on Boost.Regex, as it's a faster and more comprehensive than STD's.",
+  "main": "generate-node.js",
+  "scripts": {
+    "start": "node generate.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/TelegramMessenger/libprisma.git"
+  },
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/TelegramMessenger/libprisma/issues"
+  },
+  "homepage": "https://github.com/TelegramMessenger/libprisma#readme",
+  "dependencies": {
+    "lodash.isequal": "^4.5.0"
+  }
+}
